Updated script to auto-edit files to remove DTC dependency and to support Red-Zone no-git option

+++ handcar/regenerate_and_build.bash
@@ -14,33 +14,83 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-PAUSE='read -sn1 -p'
 
-while getopts "y" opt; do
+# SCRIPT DEFAULTS
+unset NO_GIT
+pause() {
+    #do things with parameters like $1 such as
+    echo
+    echo $1
+    echo
+}
+
+function usage {
+    echo "Usage: $(basename $0) [-in]"
+    echo "    -i  interactive mode"
+    echo "    -n  no-git mode"
+    echo ""
+    exit 1
+}
+
+while getopts ":in" opt; do
     case "$opt" in
-    y)  PAUSE='read -t5 -sn1 -p'
+    i)  # interactive
+        echo "===== option 'i' detected"
+        pause() {
+            #do things with parameters like $1 such as
+            echo
+            read -sn1 -p "$1 -- Press Enter to continue or Ctrl-C to quit"
+            echo
+        }
+        ;;
+    n)  # no git
+        echo "===== option 'n' detected"
+        NO_GIT=1
+        ;;
+    ?)  # usage
+        echo "Invalid option: -${OPTARG}"
+        echo
+        usage
         ;;
     esac
 done
 
 make clean
-rm -rf standalone
-git clone https://github.com/riscv/riscv-isa-sim standalone
-cd standalone
-git checkout 61f0dab33f7e529cc709908840311a8a7dcb23ce
 
-echo "Preparing to remove DTC dependencies"
+if [ -z "${NO_GIT}" ]; then
+    pause "Preparing to clone spike"
+    rm -rf standalone
+    git clone https://github.com/riscv/riscv-isa-sim standalone
+    cd standalone
+    git checkout 61f0dab33f7e529cc709908840311a8a7dcb23ce
+else
+    echo
+    echo ===== NO-GIT requested =====
+    echo "You are responsible for ensuring that handcar/standalone contains"
+    echo "a clone of https://github.com/riscv/riscv-isa-sim, and a checkout"
+    echo "of hash 61f0dab33f7e529cc709908840311a8a7dcb23ce"
+    echo
+    cd standalone
+fi
+
+echo "===== Preparing to remove DTC dependencies"
 
 sed -i  '/^if test x"$DTC" == xno; then :/,/fi/ s/^/# /'  ./configure
+echo
+echo "vvvvv Begin auto-edit output vvvvv"
 cat ./configure | grep -B10 -A7 'if test x"$DTC" == xno; then :'
-$PAUSE "Please review edit(s) then press Enter to continue"; echo
+echo "^^^^^ End auto-edit output ^^^^^"
+pause "===== Please review edit(s) above"
 
 sed -i  '/^  if (dtb_pid == 0)/,/^  }/ s/^/\/\/ /'             ./riscv/dts.cc
 sed -i  '/^  waitpid(dts_pid, &status, 0);/,/^  }/ s/^/\/\/ /' ./riscv/dts.cc
 sed -i  '/^  waitpid(dtb_pid, &status, 0);/,/^  }/ s/^/\/\/ /' ./riscv/dts.cc
+echo "vvvvv Begin auto-edit output vvvvv"
 cat ./riscv/dts.cc | grep -B9 -A10 -e"if (dtb_pid == 0) {" -e"waitpid(dts_pid, &status, 0);" -e"waitpid(dtb_pid, &status, 0);"
-$PAUSE "Please review edit(s) then press Enter to continue"; echo
+echo "^^^^^ End auto-edit output ^^^^^"
+pause "Please review edit(s) above"
 
+pause "===== Preparing to build spike"
 ./configure
 make -j
 cd ..
@@ -56,3 +106,11 @@ mkdir bin
 ./filesurgeon.py
 make -j8
 cp bin/handcar_cosim.so ../utils/handcar
+
+pause "===== Preparing to patch and build handcar"
+
+cd ./patcher
+./step_1_create_patches.bash
+./step_2_apply_patches.bash
+./step_3_stage_patched.bash
+cd ..
